add friend

+++ src/main/java/org/team10424102/blackserver/exceptions/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package org.team10424102.blackserver.exceptions;
 
 public class UserNotFoundException extends SystemException {
-    public UserNotFoundException(int userId) {
+    public UserNotFoundException(long userId) {
         super(String.format("找不到用户 (id = %d)", userId));
     }
 }
